Struct louds_rs::louds::Louds

pub struct Louds { /* fields omitted */ }

LOUDS (Level-Order Unary Degree Sequence).

This class can handle tree structure of virtually arbitrary number of nodes.

In fact, N (number of nodes in the tree) is designed to be limited to: N < 2^64 / 2, while each node is represented in 2bits in average.
It should be enough for almost all usecases since a binary data of length of 2^63 consumes 2^20 = 1,048,576 TB (terabytes), which is hard to handle by state-of-the-art computer architecture.

Methods

impl Louds

pub fn node_num_to_index(&self, node_num: LoudsNodeNum) -> LoudsIndex

Panics

node_num does not exist in this LOUDS.

pub fn index_to_node_num(&self, index: LoudsIndex) -> LoudsNodeNum

Panics

index does not point to any node in this LOUDS.

pub fn child_to_parent(&self, index: LoudsIndex) -> LoudsNodeNum

Panics

index does not point to any node in this LOUDS.

index == 0: (node#1 is root and doesn't have parent)

pub fn parent_to_children(&self, node_num: LoudsNodeNum) -> Vec<LoudsIndex>

Panics

node_num does not exist in this LOUDS.